requirements.txt
# Telegram Bot
python-telegram-bot==20.7
deepgram-sdk>=4.1.0
anthropic>=0.25.1
requests>=2.31.0
pathlib2>=2.3.7
python-dateutil>=2.8.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1
httpx>=0.25.2
aiofiles>=23.2.1
aiohttp>=3.9.1
Flask>=2.3.0
Werkzeug>=2.3.0
gunicorn>=21.2.0
